modularize gsm8 and mmmu test classes
